working on file structure
