+++ README.md
@@ -32,7 +32,7 @@ fbc_curation --model ./fbc_curation/examples/models/e_coli_core.xml --path ./fbc
 This creates the FBC curation files for the model in the output folder.
 
 ## Standardized reference files
-In the following the created reference output files are described and examples provided for the [`e_coli_core.xml`](fbc_curation/examples/models/e_coli_core.xml) model. All output files are tab separated files (TSV) with the first three columns being `model`, `objective`, and `status`. The column `model` encodes the SBML model id. The column `objective` encodes the SBML objective id, which is the objective which was optimized in the respective simulation. The column `status` encodes the status of the simulation. The status can be either `optimal` (optimization worked) or `infeasible` (no solution found or problem in simulation).  
+In the following the created reference output files are described and examples provided for the [`e_coli_core.xml`](examples/models/e_coli_core.xml) model. All output files are tab separated files (TSV) with the first three columns being `model`, `objective`, and `status`. The column `model` encodes the SBML model id. The column `objective` encodes the SBML objective id, which is the objective which was optimized in the respective simulation. The column `status` encodes the status of the simulation. The status can be either `optimal` (optimization worked) or `infeasible` (no solution found or problem in simulation).  
 
 ### 01 Objective value
 The objective value file `01_objective.tsv` contains the four columns with the headers `model`, `objective` and `status`, `value`. The `value` is the optimal value of the respective objective function when the model is optimized with default settings. If the status is `infeasible` the value is `NA`.
@@ -40,7 +40,7 @@ The objective value file `01_objective.tsv` contains the four columns with the h
 model	objective	status	value
 e_coli_core	obj	optimal	0.87392151
 ```
-See for instance: [`e_coli_core/01_objective.tsv`](fbc_curation/examples/results/e_coli_core/01_objective.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/simulating.html
+See for instance: [`e_coli_core/01_objective.tsv`](examples/results/e_coli_core/01_objective.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/simulating.html
 
 ### 02 Flux variability analysis (FVA)
 The FVA file `02_fva.tsv` contains five columns with the headers `model`, `objective`, `reaction`, `status`, `minimum` and `maximum`. The `minimum` and `maximum` columns contain the minimum and maximum values of the FVA. The rows are sorted based on reaction identifier. The `status` contains the status of the optimization (`optimal` or `infeasible`). If the status is `infeasible` the value is `NA`.
@@ -58,7 +58,7 @@ e_coli_core	obj	ADK1	optimal	0.0	0.0
 e_coli_core	obj	AKGDH	optimal	5.06437566	5.06437566
 ...
 ```
-See for instance: [`e_coli_core/02_fva.tsv`](fbc_curation/examples/results/e_coli_core/02_fva.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/simulating.html#Running-FVA
+See for instance: [`e_coli_core/02_fva.tsv`](examples/results/e_coli_core/02_fva.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/simulating.html#Running-FVA
 
 ### 03 Gene deletions 
 The gene deletion file `03_gene_deletion.tsv` contains five columns with the headers `model`, `objective`, `gene`, `status` and `value`. 
@@ -76,7 +76,7 @@ e_coli_core	obj	b2415	infeasible	NA
 e_coli_core	obj	b2416	infeasible	NA
 ...
 ```
-See for instance: [`e_coli_core/03_gene_deletion.tsv`](fbc_curation/examples/results/e_coli_core/03_gene_deletion.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/deletions.html
+See for instance: [`e_coli_core/03_gene_deletion.tsv`](examples/results/e_coli_core/03_gene_deletion.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/deletions.html
 
 ### 04 Reaction deletions 
 The reaction deletion file `04_reaction_deletion.tsv` contains five columns with the headers `model`, `objective`, `reaction`, `status` and `value`. 
@@ -93,7 +93,7 @@ e_coli_core	obj	ADK1	optimal	0.87392151
 e_coli_core	obj	AKGDH	optimal	0.85830741
 ...
 ```
-See for instance: [`e_coli_core/04_reaction_deletion.tsv`](fbc_curation/examples/results/e_coli_core/04_reaction_deletion.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/deletions.html
+See for instance: [`e_coli_core/04_reaction_deletion.tsv`](examples/results/e_coli_core/04_reaction_deletion.tsv). For more information: https://cobrapy.readthedocs.io/en/latest/deletions.html
 
 ## Installation
