+++ fbc_curation/__init__.py
@@ -1,4 +1,4 @@
 from pathlib import Path
+from ._version import __version__
 
-
-EXAMPLE_PATH = Path(__file__).parent / "examples"
+EXAMPLE_PATH = Path(__file__).parent.parent / "examples"
