--- fbc_curation/examples/results/e_coli_core/cameo/01_objective.tsv
@@ -1,2 +1,0 @@
-model	objective	status	value
-e_coli_core	active_objective	optimal	0.873922
