--- fbc_curation/examples/results/e_coli_core/cameo/02_fva.tsv
@@ -1,96 +1,0 @@
-model	objective	reaction	status	minimum	maximum
-e_coli_core	active_objective	ACALD	optimal	0.0	0.0
-e_coli_core	active_objective	ACALDt	optimal	0.0	0.0
-e_coli_core	active_objective	ACKr	optimal	0.0	0.0
-e_coli_core	active_objective	ACONTa	optimal	6.00725	6.00725
-e_coli_core	active_objective	ACONTb	optimal	6.00725	6.00725
-e_coli_core	active_objective	ACt2r	optimal	0.0	0.0
-e_coli_core	active_objective	ADK1	optimal	0.0	0.0
-e_coli_core	active_objective	AKGDH	optimal	5.064376	5.064376
-e_coli_core	active_objective	AKGt2r	optimal	0.0	0.0
-e_coli_core	active_objective	ALCD2x	optimal	0.0	0.0
-e_coli_core	active_objective	ATPM	optimal	8.39	8.39
-e_coli_core	active_objective	ATPS4r	optimal	45.51401	45.51401
-e_coli_core	active_objective	BIOMASS_Ecoli_core_w_GAM	optimal	0.873922	0.873922
-e_coli_core	active_objective	CO2t	optimal	22.809833	22.809833
-e_coli_core	active_objective	CS	optimal	6.00725	6.00725
-e_coli_core	active_objective	CYTBD	optimal	43.598985	43.598985
-e_coli_core	active_objective	D_LACt2	optimal	0.0	0.0
-e_coli_core	active_objective	ENO	optimal	14.71614	14.71614
-e_coli_core	active_objective	ETOHt2r	optimal	0.0	0.0
-e_coli_core	active_objective	EX_ac_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_acald_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_akg_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_co2_e	optimal	22.809833	22.809833
-e_coli_core	active_objective	EX_etoh_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_for_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_fru_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_fum_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_glc__D_e	optimal	10.0	10.0
-e_coli_core	active_objective	EX_gln__L_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_glu__L_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_h2o_e	optimal	29.175827	29.175827
-e_coli_core	active_objective	EX_h_e	optimal	17.530865	17.530865
-e_coli_core	active_objective	EX_lac__D_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_mal__L_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_nh4_e	optimal	4.765319	4.765319
-e_coli_core	active_objective	EX_o2_e	optimal	21.799493	21.799493
-e_coli_core	active_objective	EX_pi_e	optimal	3.214895	3.214895
-e_coli_core	active_objective	EX_pyr_e	optimal	0.0	0.0
-e_coli_core	active_objective	EX_succ_e	optimal	0.0	0.0
-e_coli_core	active_objective	FBA	optimal	7.477382	7.477382
-e_coli_core	active_objective	FBP	optimal	0.0	0.0
-e_coli_core	active_objective	FORt	optimal	0.0	0.0
-e_coli_core	active_objective	FORt2	optimal	0.0	0.0
-e_coli_core	active_objective	FRD7	optimal	0.0	994.935624
-e_coli_core	active_objective	FRUpts2	optimal	0.0	0.0
-e_coli_core	active_objective	FUM	optimal	5.064376	5.064376
-e_coli_core	active_objective	FUMt2_2	optimal	0.0	0.0
-e_coli_core	active_objective	G6PDH2r	optimal	4.959985	4.959985
-e_coli_core	active_objective	GAPD	optimal	16.023526	16.023526
-e_coli_core	active_objective	GLCpts	optimal	10.0	10.0
-e_coli_core	active_objective	GLNS	optimal	0.223462	0.223462
-e_coli_core	active_objective	GLNabc	optimal	0.0	0.0
-e_coli_core	active_objective	GLUDy	optimal	4.541857	4.541857
-e_coli_core	active_objective	GLUN	optimal	0.0	0.0
-e_coli_core	active_objective	GLUSy	optimal	0.0	0.0
-e_coli_core	active_objective	GLUt2r	optimal	0.0	0.0
-e_coli_core	active_objective	GND	optimal	4.959985	4.959985
-e_coli_core	active_objective	H2Ot	optimal	29.175827	29.175827
-e_coli_core	active_objective	ICDHyr	optimal	6.00725	6.00725
-e_coli_core	active_objective	ICL	optimal	0.0	0.0
-e_coli_core	active_objective	LDH_D	optimal	0.0	0.0
-e_coli_core	active_objective	MALS	optimal	0.0	0.0
-e_coli_core	active_objective	MALt2_2	optimal	0.0	0.0
-e_coli_core	active_objective	MDH	optimal	5.064376	5.064376
-e_coli_core	active_objective	ME1	optimal	0.0	0.0
-e_coli_core	active_objective	ME2	optimal	0.0	0.0
-e_coli_core	active_objective	NADH16	optimal	38.53461	38.53461
-e_coli_core	active_objective	NADTRHD	optimal	0.0	0.0
-e_coli_core	active_objective	NH4t	optimal	4.765319	4.765319
-e_coli_core	active_objective	O2t	optimal	21.799493	21.799493
-e_coli_core	active_objective	PDH	optimal	9.282533	9.282533
-e_coli_core	active_objective	PFK	optimal	7.477382	7.477382
-e_coli_core	active_objective	PFL	optimal	0.0	0.0
-e_coli_core	active_objective	PGI	optimal	4.860861	4.860861
-e_coli_core	active_objective	PGK	optimal	16.023526	16.023526
-e_coli_core	active_objective	PGL	optimal	4.959985	4.959985
-e_coli_core	active_objective	PGM	optimal	14.71614	14.71614
-e_coli_core	active_objective	PIt2r	optimal	3.214895	3.214895
-e_coli_core	active_objective	PPC	optimal	2.504309	2.504309
-e_coli_core	active_objective	PPCK	optimal	0.0	0.0
-e_coli_core	active_objective	PPS	optimal	0.0	0.0
-e_coli_core	active_objective	PTAr	optimal	0.0	0.0
-e_coli_core	active_objective	PYK	optimal	1.758177	1.758177
-e_coli_core	active_objective	PYRt2	optimal	0.0	0.0
-e_coli_core	active_objective	RPE	optimal	2.678482	2.678482
-e_coli_core	active_objective	RPI	optimal	2.281503	2.281503
-e_coli_core	active_objective	SUCCt2_2	optimal	0.0	0.0
-e_coli_core	active_objective	SUCCt3	optimal	0.0	0.0
-e_coli_core	active_objective	SUCDi	optimal	5.064376	1000.0
-e_coli_core	active_objective	SUCOAS	optimal	5.064376	5.064376
-e_coli_core	active_objective	TALA	optimal	1.496984	1.496984
-e_coli_core	active_objective	THD2	optimal	0.0	0.0
-e_coli_core	active_objective	TKT1	optimal	1.496984	1.496984
-e_coli_core	active_objective	TKT2	optimal	1.181498	1.181498
-e_coli_core	active_objective	TPI	optimal	7.477382	7.477382
