--- fbc_curation/examples/results/e_coli_core/cobrapy/03_gene_deletion.tsv
@@ -1,138 +1,0 @@
-model	objective	gene	status	value
-e_coli_core	active_objective	b0008	optimal	0.873922
-e_coli_core	active_objective	b0114	optimal	0.796696
-e_coli_core	active_objective	b0115	optimal	0.796696
-e_coli_core	active_objective	b0116	optimal	0.782351
-e_coli_core	active_objective	b0118	optimal	0.873922
-e_coli_core	active_objective	b0351	optimal	0.873922
-e_coli_core	active_objective	b0356	optimal	0.873922
-e_coli_core	active_objective	b0451	optimal	0.873922
-e_coli_core	active_objective	b0474	optimal	0.873922
-e_coli_core	active_objective	b0485	optimal	0.873922
-e_coli_core	active_objective	b0720	optimal	0.0
-e_coli_core	active_objective	b0721	optimal	0.814298
-e_coli_core	active_objective	b0722	optimal	0.814298
-e_coli_core	active_objective	b0723	optimal	0.814298
-e_coli_core	active_objective	b0724	optimal	0.814298
-e_coli_core	active_objective	b0726	optimal	0.858307
-e_coli_core	active_objective	b0727	optimal	0.858307
-e_coli_core	active_objective	b0728	optimal	0.858307
-e_coli_core	active_objective	b0729	optimal	0.858307
-e_coli_core	active_objective	b0733	optimal	0.873922
-e_coli_core	active_objective	b0734	optimal	0.873922
-e_coli_core	active_objective	b0755	optimal	0.873922
-e_coli_core	active_objective	b0767	optimal	0.863813
-e_coli_core	active_objective	b0809	optimal	0.873922
-e_coli_core	active_objective	b0810	optimal	0.873922
-e_coli_core	active_objective	b0811	optimal	0.873922
-e_coli_core	active_objective	b0875	optimal	0.873922
-e_coli_core	active_objective	b0902	optimal	0.873922
-e_coli_core	active_objective	b0903	optimal	0.873922
-e_coli_core	active_objective	b0904	optimal	0.873922
-e_coli_core	active_objective	b0978	optimal	0.873922
-e_coli_core	active_objective	b0979	optimal	0.873922
-e_coli_core	active_objective	b1101	optimal	0.873922
-e_coli_core	active_objective	b1136	optimal	0.0
-e_coli_core	active_objective	b1241	optimal	0.873922
-e_coli_core	active_objective	b1276	optimal	0.873922
-e_coli_core	active_objective	b1297	optimal	0.873922
-e_coli_core	active_objective	b1380	optimal	0.873922
-e_coli_core	active_objective	b1478	optimal	0.873922
-e_coli_core	active_objective	b1479	optimal	0.873922
-e_coli_core	active_objective	b1524	optimal	0.873922
-e_coli_core	active_objective	b1602	optimal	0.873922
-e_coli_core	active_objective	b1603	optimal	0.873922
-e_coli_core	active_objective	b1611	optimal	0.873922
-e_coli_core	active_objective	b1612	optimal	0.873922
-e_coli_core	active_objective	b1621	optimal	0.873922
-e_coli_core	active_objective	b1676	optimal	0.873922
-e_coli_core	active_objective	b1702	optimal	0.873922
-e_coli_core	active_objective	b1723	optimal	0.873922
-e_coli_core	active_objective	b1761	optimal	0.851389
-e_coli_core	active_objective	b1773	optimal	0.873922
-e_coli_core	active_objective	b1779	optimal	0.0
-e_coli_core	active_objective	b1812	optimal	0.873922
-e_coli_core	active_objective	b1817	optimal	0.873922
-e_coli_core	active_objective	b1818	optimal	0.873922
-e_coli_core	active_objective	b1819	optimal	0.873922
-e_coli_core	active_objective	b1849	optimal	0.873922
-e_coli_core	active_objective	b1852	optimal	0.863813
-e_coli_core	active_objective	b1854	optimal	0.873922
-e_coli_core	active_objective	b2029	optimal	0.863813
-e_coli_core	active_objective	b2097	optimal	0.873922
-e_coli_core	active_objective	b2133	optimal	0.873922
-e_coli_core	active_objective	b2276	optimal	0.211663
-e_coli_core	active_objective	b2277	optimal	0.211663
-e_coli_core	active_objective	b2278	optimal	0.211663
-e_coli_core	active_objective	b2279	optimal	0.211663
-e_coli_core	active_objective	b2280	optimal	0.211663
-e_coli_core	active_objective	b2281	optimal	0.211663
-e_coli_core	active_objective	b2282	optimal	0.211663
-e_coli_core	active_objective	b2283	optimal	0.211663
-e_coli_core	active_objective	b2284	optimal	0.211663
-e_coli_core	active_objective	b2285	optimal	0.211663
-e_coli_core	active_objective	b2286	optimal	0.211663
-e_coli_core	active_objective	b2287	optimal	0.211663
-e_coli_core	active_objective	b2288	optimal	0.211663
-e_coli_core	active_objective	b2296	optimal	0.873922
-e_coli_core	active_objective	b2297	optimal	0.873922
-e_coli_core	active_objective	b2415	infeasible	
-e_coli_core	active_objective	b2416	infeasible	
-e_coli_core	active_objective	b2417	optimal	0.873922
-e_coli_core	active_objective	b2458	optimal	0.873922
-e_coli_core	active_objective	b2463	optimal	0.873922
-e_coli_core	active_objective	b2464	optimal	0.873922
-e_coli_core	active_objective	b2465	optimal	0.873922
-e_coli_core	active_objective	b2492	optimal	0.873922
-e_coli_core	active_objective	b2579	optimal	0.873922
-e_coli_core	active_objective	b2587	optimal	0.873922
-e_coli_core	active_objective	b2779	optimal	0.0
-e_coli_core	active_objective	b2914	optimal	0.873922
-e_coli_core	active_objective	b2925	optimal	0.873922
-e_coli_core	active_objective	b2926	optimal	0.0
-e_coli_core	active_objective	b2935	optimal	0.873922
-e_coli_core	active_objective	b2975	optimal	0.873922
-e_coli_core	active_objective	b2976	optimal	0.873922
-e_coli_core	active_objective	b2987	optimal	0.873922
-e_coli_core	active_objective	b3114	optimal	0.873922
-e_coli_core	active_objective	b3115	optimal	0.873922
-e_coli_core	active_objective	b3212	optimal	0.873922
-e_coli_core	active_objective	b3213	optimal	0.873922
-e_coli_core	active_objective	b3236	optimal	0.825819
-e_coli_core	active_objective	b3386	optimal	0.873922
-e_coli_core	active_objective	b3403	optimal	0.873922
-e_coli_core	active_objective	b3493	optimal	0.873922
-e_coli_core	active_objective	b3528	optimal	0.873922
-e_coli_core	active_objective	b3603	optimal	0.873922
-e_coli_core	active_objective	b3612	optimal	0.873922
-e_coli_core	active_objective	b3731	optimal	0.37423
-e_coli_core	active_objective	b3732	optimal	0.37423
-e_coli_core	active_objective	b3733	optimal	0.37423
-e_coli_core	active_objective	b3734	optimal	0.37423
-e_coli_core	active_objective	b3735	optimal	0.37423
-e_coli_core	active_objective	b3736	optimal	0.37423
-e_coli_core	active_objective	b3737	optimal	0.37423
-e_coli_core	active_objective	b3738	optimal	0.37423
-e_coli_core	active_objective	b3739	optimal	0.873922
-e_coli_core	active_objective	b3870	optimal	0.873922
-e_coli_core	active_objective	b3916	optimal	0.873922
-e_coli_core	active_objective	b3919	optimal	0.704037
-e_coli_core	active_objective	b3925	optimal	0.873922
-e_coli_core	active_objective	b3951	optimal	0.873922
-e_coli_core	active_objective	b3952	optimal	0.873922
-e_coli_core	active_objective	b3956	optimal	0.870745
-e_coli_core	active_objective	b3962	optimal	0.873922
-e_coli_core	active_objective	b4014	optimal	0.873922
-e_coli_core	active_objective	b4015	optimal	0.873922
-e_coli_core	active_objective	b4025	optimal	0.86316
-e_coli_core	active_objective	b4077	optimal	0.873922
-e_coli_core	active_objective	b4090	optimal	0.873922
-e_coli_core	active_objective	b4122	optimal	0.873922
-e_coli_core	active_objective	b4151	optimal	0.873922
-e_coli_core	active_objective	b4152	optimal	0.873922
-e_coli_core	active_objective	b4153	optimal	0.873922
-e_coli_core	active_objective	b4154	optimal	0.873922
-e_coli_core	active_objective	b4232	optimal	0.873922
-e_coli_core	active_objective	b4301	optimal	0.873922
-e_coli_core	active_objective	b4395	optimal	0.873922
-e_coli_core	active_objective	s0001	optimal	0.211141
