--- fbc_curation/examples/results/e_coli_core/cobrapy/04_reaction_deletion.tsv
@@ -1,96 +1,0 @@
-model	objective	reaction	status	value
-e_coli_core	active_objective	ACALD	optimal	0.873922
-e_coli_core	active_objective	ACALDt	optimal	0.873922
-e_coli_core	active_objective	ACKr	optimal	0.873922
-e_coli_core	active_objective	ACONTa	optimal	0.0
-e_coli_core	active_objective	ACONTb	optimal	0.0
-e_coli_core	active_objective	ACt2r	optimal	0.873922
-e_coli_core	active_objective	ADK1	optimal	0.873922
-e_coli_core	active_objective	AKGDH	optimal	0.858307
-e_coli_core	active_objective	AKGt2r	optimal	0.873922
-e_coli_core	active_objective	ALCD2x	optimal	0.873922
-e_coli_core	active_objective	ATPM	optimal	0.916647
-e_coli_core	active_objective	ATPS4r	optimal	0.37423
-e_coli_core	active_objective	BIOMASS_Ecoli_core_w_GAM	optimal	0.0
-e_coli_core	active_objective	CO2t	optimal	0.46167
-e_coli_core	active_objective	CS	optimal	0.0
-e_coli_core	active_objective	CYTBD	optimal	0.211663
-e_coli_core	active_objective	D_LACt2	optimal	0.873922
-e_coli_core	active_objective	ENO	optimal	0.0
-e_coli_core	active_objective	ETOHt2r	optimal	0.873922
-e_coli_core	active_objective	EX_ac_e	optimal	0.873922
-e_coli_core	active_objective	EX_acald_e	optimal	0.873922
-e_coli_core	active_objective	EX_akg_e	optimal	0.873922
-e_coli_core	active_objective	EX_co2_e	optimal	0.46167
-e_coli_core	active_objective	EX_etoh_e	optimal	0.873922
-e_coli_core	active_objective	EX_for_e	optimal	0.873922
-e_coli_core	active_objective	EX_fru_e	optimal	0.873922
-e_coli_core	active_objective	EX_fum_e	optimal	0.873922
-e_coli_core	active_objective	EX_glc__D_e	infeasible	
-e_coli_core	active_objective	EX_gln__L_e	optimal	0.873922
-e_coli_core	active_objective	EX_glu__L_e	optimal	0.873922
-e_coli_core	active_objective	EX_h2o_e	optimal	0.383515
-e_coli_core	active_objective	EX_h_e	optimal	0.0
-e_coli_core	active_objective	EX_lac__D_e	optimal	0.873922
-e_coli_core	active_objective	EX_mal__L_e	optimal	0.873922
-e_coli_core	active_objective	EX_nh4_e	optimal	0.0
-e_coli_core	active_objective	EX_o2_e	optimal	0.211663
-e_coli_core	active_objective	EX_pi_e	optimal	0.0
-e_coli_core	active_objective	EX_pyr_e	optimal	0.873922
-e_coli_core	active_objective	EX_succ_e	optimal	0.873922
-e_coli_core	active_objective	FBA	optimal	0.704037
-e_coli_core	active_objective	FBP	optimal	0.873922
-e_coli_core	active_objective	FORt	optimal	0.873922
-e_coli_core	active_objective	FORt2	optimal	0.873922
-e_coli_core	active_objective	FRD7	optimal	0.873922
-e_coli_core	active_objective	FRUpts2	optimal	0.873922
-e_coli_core	active_objective	FUM	optimal	0.814298
-e_coli_core	active_objective	FUMt2_2	optimal	0.873922
-e_coli_core	active_objective	G6PDH2r	optimal	0.863813
-e_coli_core	active_objective	GAPD	optimal	0.0
-e_coli_core	active_objective	GLCpts	infeasible	
-e_coli_core	active_objective	GLNS	optimal	0.0
-e_coli_core	active_objective	GLNabc	optimal	0.873922
-e_coli_core	active_objective	GLUDy	optimal	0.851389
-e_coli_core	active_objective	GLUN	optimal	0.873922
-e_coli_core	active_objective	GLUSy	optimal	0.873922
-e_coli_core	active_objective	GLUt2r	optimal	0.873922
-e_coli_core	active_objective	GND	optimal	0.863813
-e_coli_core	active_objective	H2Ot	optimal	0.383515
-e_coli_core	active_objective	ICDHyr	optimal	0.0
-e_coli_core	active_objective	ICL	optimal	0.873922
-e_coli_core	active_objective	LDH_D	optimal	0.873922
-e_coli_core	active_objective	MALS	optimal	0.873922
-e_coli_core	active_objective	MALt2_2	optimal	0.873922
-e_coli_core	active_objective	MDH	optimal	0.825819
-e_coli_core	active_objective	ME1	optimal	0.873922
-e_coli_core	active_objective	ME2	optimal	0.873922
-e_coli_core	active_objective	NADH16	optimal	0.211663
-e_coli_core	active_objective	NADTRHD	optimal	0.873922
-e_coli_core	active_objective	NH4t	optimal	0.0
-e_coli_core	active_objective	O2t	optimal	0.211663
-e_coli_core	active_objective	PDH	optimal	0.796696
-e_coli_core	active_objective	PFK	optimal	0.704037
-e_coli_core	active_objective	PFL	optimal	0.873922
-e_coli_core	active_objective	PGI	optimal	0.86316
-e_coli_core	active_objective	PGK	optimal	0.0
-e_coli_core	active_objective	PGL	optimal	0.863813
-e_coli_core	active_objective	PGM	optimal	0.0
-e_coli_core	active_objective	PIt2r	optimal	0.0
-e_coli_core	active_objective	PPC	optimal	0.870745
-e_coli_core	active_objective	PPCK	optimal	0.873922
-e_coli_core	active_objective	PPS	optimal	0.873922
-e_coli_core	active_objective	PTAr	optimal	0.873922
-e_coli_core	active_objective	PYK	optimal	0.864926
-e_coli_core	active_objective	PYRt2	optimal	0.873922
-e_coli_core	active_objective	RPE	optimal	0.865716
-e_coli_core	active_objective	RPI	optimal	0.0
-e_coli_core	active_objective	SUCCt2_2	optimal	0.873922
-e_coli_core	active_objective	SUCCt3	optimal	0.873922
-e_coli_core	active_objective	SUCDi	optimal	0.814298
-e_coli_core	active_objective	SUCOAS	optimal	0.858307
-e_coli_core	active_objective	TALA	optimal	0.864759
-e_coli_core	active_objective	THD2	optimal	0.873922
-e_coli_core	active_objective	TKT1	optimal	0.864759
-e_coli_core	active_objective	TKT2	optimal	0.866674
-e_coli_core	active_objective	TPI	optimal	0.704037
