+++ fbc_curation/tests/__init__.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+PATH_TEST_DATA = Path(__file__).parent / "data"
