+++ fbc_curation/tests/data/outputs/e_coli_core/cameo/01_objective.tsv
@@ -1,0 +1,2 @@
+model	objective	status	value
+e_coli_core	active_objective	optimal	0.873922
